+++ .gitignore
@@ -14,7 +14,7 @@
 # Binaries
 cmd/shortener/main
 cmd/shortener/shortener
-bin
+bin/
 
 # Dependency directories (remove the comment below to include it)
 vendor/
