+++ internal/app/helpers_test.go
@@ -1,8 +1,9 @@
 package app
 
 import (
-	"github.com/stretchr/testify/assert"
 	"testing"
+
+	"github.com/stretchr/testify/assert"
 )
 
 func Test_generateID(t *testing.T) {
